Create common abstraction for main_menu and tab items.
