+++ app/models/spree/admin/main_menu/root.rb
@@ -2,6 +2,8 @@ module Spree
   module Admin
     module MainMenu
       class Root < Section
+        include ::Spree::Admin::ItemManager
+
         attr_reader :items
 
         def initialize
