+++ app/models/spree/admin/tabs/root.rb
@@ -2,22 +2,12 @@ module Spree
   module Admin
     module Tabs
       class Root
+        include ::Spree::Admin::ItemManager
+
         attr_reader :items
 
         def initialize
           @items = []
-        end
-
-        def add(item)
-          raise KeyError, "Item with key #{item.key} already exists" if index_for_key(item.key)
-
-          @items << item
-        end
-
-        private
-
-        def index_for_key(key)
-          @items.index { |e| e.key == key }
         end
       end
     end
